+++ src/settings/SettingsModal.js
@@ -7,7 +7,7 @@ import Toggle from 'react-toggle'
 import '../App.css';
 import './settings.css';
 import PreferencesContext from '../contexts/PreferencesContext';
-import ConfigurationContext from '../contexts/ConfigurationContext';
+import ConfigurationContext from '../configuration/ConfigurationContext';
 import { SUPPORTED_CARDS, APP } from '../Constants'
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
